RF: Make obsolete resolve_path() internal helper of annotate_paths()
That is its last user. Also relocate the associated tests.

Ping gh-3778
